src/main/java/com/coyoapp/tinytask/service/DefaultUserService.java
package com.coyoapp.tinytask.service;

import com.coyoapp.tinytask.domain.User;
import com.coyoapp.tinytask.dto.UserRequest;
import com.coyoapp.tinytask.dto.UserResponse;
import com.coyoapp.tinytask.repository.UserRepository;
import lombok.RequiredArgsConstructor;
import lombok.extern.slf4j.Slf4j;
import ma.glasnost.orika.MapperFacade;
import org.springframework.stereotype.Component;
import org.springframework.transaction.annotation.Transactional;

@Slf4j
@Component
@RequiredArgsConstructor
public class DefaultUserService implements UserService {

  private final UserRepository userRepository;
  private final MapperFacade mapperFacade;

  @Override
  @Transactional
  public UserResponse createUser(UserRequest userRequest) {
    log.debug("createUser(userRequest={})", userRequest);
    User user = mapperFacade.map(userRequest,User.class);
    return transformToResponse(userRepository.save(user));
  }

  private UserResponse transformToResponse(User user) {
    return mapperFacade.map(user, UserResponse.class);
  }

}
